requirements.txt
# Core simulation dependencies
gymnasium==1.0.0
matplotlib==3.5.1
numpy==1.21.5
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
PyYAML==6.0.2
scipy==1.8.0
setuptools==69.5.1
scikit-image==0.20.0
vispy==0.13.0

